ProjectTests: common fixtures.
