+++ pytest.ini
@@ -5,6 +5,7 @@ markers =
     config
     request
     phantom
+    armory
 flake8-ignore =
     E501
     .git/*.* ALL
